Move `CreateSelectColumnsWithMeasuresAggregated` to a separate file.
